# openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Management System API
  description: |
    Multi-client, multi-strategy investment portfolio management backend.

    **Hierarchy:** Client → Account → Sleeve (Account × Strategy join) → Trades/Positions

    - A **Strategy** is a firm-wide definition (e.g. "Growth", "Income") — global and reusable.
    - A **Sleeve** is one strategy as run in one account — it owns trades, positions, and sync configs.
    - All resource IDs are UUIDs (string).
  version: 1.0.0
  contact:
    email: venkatasudheerc@gmail.com
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/v1/clients/{client_id}/trades:
    get:
      tags: [clients]
      summary: All trades across all sleeves for a client
      operationId: getClientTrades
      parameters:
        - $ref: "#/components/parameters/client_id"
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Paginated trade list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientTradesResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Accounts ───────────────────────────────────────────────────────────────
...
  /api/v1/clients/{client_id}/accounts/{account_id}/sleeves/{sleeve_id}/trades:
    get:
      tags: [sleeves]
      summary: Trades for a sleeve (paginated)
      operationId: getSleeveTrades
      parameters:
        - $ref: "#/components/parameters/client_id"
        - $ref: "#/components/parameters/account_id"
        - $ref: "#/components/parameters/sleeve_id"
        - $ref: "#/components/parameters/skip"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Paginated trade list for the sleeve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleeveTradesResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/strategies/:
    post:
      tags: [strategies]
      summary: Create a global strategy definition
      description: Strategy names are unique case-insensitively. Returns 409 on duplicate.
      operationId: createStrategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyCreate"
      responses:
        "201":
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyResponse"
        "409":
          description: A strategy with this name already exists (case-insensitive)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      tags: [strategies]
      summary: List all global strategy definitions
      operationId: listStrategies
      responses:
        "200":
          description: List of strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StrategyResponse"
  /api/v1/strategies/{strategy_id}:
    get:
      tags: [strategies]
      summary: Get a strategy definition by ID
      operationId: getStrategy
      parameters:
        - $ref: "#/components/parameters/strategy_id"
      responses:
        "200":
          description: Strategy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/strategies/{strategy_id}/performance:
    get:
      tags: [strategies]
      summary: Strategy performance roll-up across all accounts running it
      operationId: getStrategyPerformance
      parameters:
        - $ref: "#/components/parameters/strategy_id"
        - $ref: "#/components/parameters/start_date"
        - $ref: "#/components/parameters/end_date"
      responses:
        "200":
          description: Strategy performance with per-sleeve children
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LevelPerformance"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/strategies/{strategy_id}/performance/monthly:
    get:
      tags: [strategies]
      summary: Strategy month-on-month returns since inception
      operationId: getStrategyMonthlyPerformance
      parameters:
        - $ref: "#/components/parameters/strategy_id"
      responses:
        "200":
          description: Monthly return breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReturnsResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/strategies/{strategy_id}/performance/returns-series:
    get:
      tags: [strategies]
      summary: Weekly TWR + MWR timeseries for a strategy
      operationId: getStrategyReturnsSeries
      parameters:
        - $ref: "#/components/parameters/strategy_id"
        - name: start_date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Weekly returns series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnsSeriesResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/trades/:
    post:
      tags: [trades]
      summary: Create a new trade
      operationId: createTrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeCreate"
      responses:
        "201":
          description: Trade created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeResponse"
        "404":
          description: Sleeve not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ─── Admin ──────────────────────────────────────────────────────────────────
...
components:
  parameters:
    client_id:
      name: client_id
      in: path
      required: true
      schema:
        type: string
      description: Client UUID
    account_id:
      name: account_id
      in: path
      required: true
      schema:
        type: string
      description: Account UUID
    sleeve_id:
      name: sleeve_id
      in: path
      required: true
      schema:
        type: string
      description: Sleeve UUID
    strategy_id:
      name: strategy_id
      in: path
      required: true
      schema:
        type: string
      description: Strategy UUID
    start_date:
      name: start_date
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Filter start date (YYYY-MM-DD)
    end_date:
      name: end_date
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Filter end date (YYYY-MM-DD)
    skip:
      name: skip
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Pagination offset
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      description: Maximum number of results to return
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Request body failed Pydantic validation
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items: {}
                    msg:
                      type: string
                    type:
                      type: string
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
      required: [detail]

    # ── Clients ──────────────────────────────────────────────────────────────
...
    ClientTradesResponse:
      type: object
      properties:
        client_id:
          type: string
        total:
          type: integer
        skip:
          type: integer
        limit:
          type: integer
        trades:
          type: array
          items:
            $ref: "#/components/schemas/TradeSummary"
    TradeSummary:
      type: object
      properties:
        id:
          type: string
        sleeve_id:
          type: string
        trade_date:
          type: string
          format: date
        symbol:
          type: string
        action:
          $ref: "#/components/schemas/TradeAction"
        quantity:
          type: number
        price:
          type: number
        commission:
          type: number
        notes:
          type: string

    # ── Accounts ─────────────────────────────────────────────────────────────
...
    SleeveTradesResponse:
      type: object
      properties:
        sleeve_id:
          type: string
        total:
          type: integer
        trades:
          type: array
          items:
            $ref: "#/components/schemas/TradeSummary"

    # ── Strategies ───────────────────────────────────────────────────────────
    StrategyCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Growth
        description:
          type: string
          default: ""
    StrategyResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true

    # ── Trades ───────────────────────────────────────────────────────────────
    TradeAction:
      type: string
      enum: [BUY, SELL]
    TradeCreate:
      type: object
      required: [sleeve_id, trade_date, symbol, action, quantity, price]
      properties:
        sleeve_id:
          type: string
        trade_date:
          type: string
          format: date
          example: "2024-01-15"
        symbol:
          type: string
          example: AAPL
        action:
          $ref: "#/components/schemas/TradeAction"
        quantity:
          type: number
          exclusiveMinimum: 0
          example: 100
        price:
          type: number
          exclusiveMinimum: 0
          example: 185.50
        commission:
          type: number
          minimum: 0
          default: 0
        notes:
          type: string
          default: ""
    TradeResponse:
      type: object
      properties:
        id:
          type: string
        sleeve_id:
          type: string
        trade_date:
          type: string
          format: date
        symbol:
          type: string
        action:
          $ref: "#/components/schemas/TradeAction"
        quantity:
          type: number
        price:
          type: number
        commission:
          type: number
        notes:
          type: string
        created_at:
          type: string
          format: date-time

    # ── Positions ────────────────────────────────────────────────────────────
...
    PerformancePoint:
      type: object
      properties:
        date:
          type: string
          format: date
        value:
          type: number
          nullable: true
    PerformanceSummary:
      type: object
      properties:
        mwr_pct:
          type: number
          nullable: true
          description: Money Weighted Return (IRR) as a decimal (e.g. 0.12 = 12%)
        twr_pct:
          type: number
          nullable: true
          description: Time Weighted Return (geometric linking) as a decimal
        twr_note:
          type: string
          nullable: true
        total_return_pct:
          type: number
          nullable: true
        total_cost_basis:
          type: number
        total_market_value:
          type: number
        total_unrealized_gain:
          type: number
        total_invested:
          type: number
        total_proceeds:
          type: number
        total_realized_gain:
          type: number
        closed_positions_count:
          type: integer
        trades_count:
          type: integer
    ChildPerformance:
      type: object
      properties:
        level:
          type: string
        id:
          type: string
        name:
          type: string
          nullable: true
        summary:
          $ref: "#/components/schemas/PerformanceSummary"
    LevelPerformance:
      type: object
      properties:
        level:
          type: string
          enum: [client, account, sleeve, strategy]
        id:
          type: string
        name:
          type: string
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        summary:
          $ref: "#/components/schemas/PerformanceSummary"
        timeseries:
          type: array
          items:
            $ref: "#/components/schemas/PerformancePoint"
        children:
          type: array
          items:
            $ref: "#/components/schemas/ChildPerformance"
    MonthlyReturn:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        month_label:
          type: string
        start_value:
          type: number
        end_value:
          type: number
        cash_in:
          type: number
        cash_out:
          type: number
        net_cash_flow:
          type: number
        realized_gain:
          type: number
        return_pct:
          type: number
          nullable: true
          description: Monthly return as a decimal (e.g. 0.034 = 3.4%); null for inception month
    MonthlyReturnsResponse:
      type: object
      properties:
        months:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyReturn"
        cumulative_return_pct:
          type: number
          nullable: true

    # ── Weekly Returns Series ────────────────────────────────────────────────
    WeeklyReturnPoint:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Monday of the snapshot week
        twr_cumul:
          type: number
          nullable: true
          description: Cumulative TWR from inception to this week (decimal, e.g. 0.05 = 5%)
        mwr_cumul:
          type: number
          nullable: true
          description: Cumulative MWR (IRR holding-period) from inception to this week
    ReturnsSeriesResponse:
      type: object
      properties:
        level:
          type: string
          enum: [client, account, sleeve, strategy]
        id:
          type: string
        series:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyReturnPoint"

    # ── Market Prices ────────────────────────────────────────────────────────
...